New HIT Type

New HIT Type

Usage

new_hit_type(
  title = "eviction-address-transcription",
  description = "Find and transcribe the DEFENDENT'S address from a court document pdf",
  reward = "0.15",
  duration = pyMTurkR::seconds(minutes = 10),
  keywords = "address, text, transcribe, entry, data",
  auto.approval.delay = pyMTurkR::seconds(days = 3),
  ...
)

Arguments

title

The title for the HIT Type

description

A description to be shown to workers

reward

A string representing amount to be paid for successful completion in USD

duration

The time in seconds that a worker has to complete the assignment

keywords

A string with comma separated terms to be used to search for HITs of this type

auto.approval.delay

The time in seconds after which a completed assignment is auto-approved

...

Additional arguments passed to pyMTurkR::CreateHITType()

Value

The HIT Type id

Examples

if (FALSE) {
new_hit_type()
new_hit_type(reward = "0.20")
new_hit_type(duration = pyMTurkR::seconds(minutes = 20))
}